backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
sse-starlette==2.1.3

# AI / LLM
langchain==1.2.9
langchain-core==1.2.9
langchain-google-genai==4.2.0
langchain-ollama==1.0.1
langchain-openai==1.1.7
langgraph==1.0.8

# Database
supabase==2.11.0
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.11
asyncpg==0.29.0  # Async PostgreSQL driver

# Flight Data
amadeus==12.0.0

# Google Calendar
google-api-python-client==2.157.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Email
resend==2.21.0

# Settings & Validation
pydantic>=2.10.4
pydantic-settings>=2.7.1
email-validator>=2.1.0

# Security
PyJWT==2.10.1
bcrypt==4.0.1  # Compatible version, or use bcrypt>=4.1.0 with direct usage

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
python-dateutil==2.9.0
